requirements.txt
playwright==1.49.1
playwright-stealth==1.0.6
openai==1.58.0
google-generativeai==0.8.0
python-dotenv==1.0.1
Pillow==11.0.0
flask==3.1.0
